ARCH-1: extract properties/tags dispatch arms (iter-225)
